refactor(server): predicate-shape.test.ts is about something else, so the module needs another name
`predicate-shape.test.ts` already exists and tests near-miss predicate rejections, importing
from predicate-parse and predicate-eval. A new `predicate-shape.ts` sitting beside it reads as
that test's subject and is not, which is a half-hour somebody loses later.

Renamed to predicate-asks.ts, which says what it answers: what does this predicate ask about.

Also records why it is not in predicate-schema.ts next to `predicateFieldsFor`, where a reader
would reasonably expect it: five open contributor PRs are queued on that file and on
predicate-eval.ts, and a one-function import is not worth handing five people a conflict.
